[refactor] tests/local_tests: organize by model family
Reorganize tests/local_tests/ from bucket-based subdirs (encoders/,
pipelines/, transformers/, vaes/) to one directory per model family
(<family>/...), matching the layout produced by the add-model-prep
skill. The previous bucket layout split a single port across four
parallel directories, making it hard to see all parity coverage for a
given family at a glance and to attach per-family setup docs.

New layout:

  tests/local_tests/<family>/
  +- README.md         # reference assets, setup, test inventory
  +- PORT_STATUS.md    # phase, component matrix, parity commands
  +- __init__.py
  +- test_<family>_<component>_*.py

Families introduced (matching the existing test inventory):
  - gamecraft, gen3c, kandinsky5, ltx2, sd35 (already family-laid-out),
    stable_audio, wan22

Other changes required by the move:
  - stable_audio tests now use a relative import for _stable_audio_helpers
    (which moved into the family dir).
  - wan22/test_i2v_record_no_clip.py: bumped repo-root resolution by one
    level to match the deeper nesting; refreshed the now-stale comment
    about transformers-package shadowing (the bucket no longer exists).
  - Pytest invocation paths in test docstrings updated to the new
    locations.
  - tests/README.md, tests/local_tests/README.md, and
    docs/contributing/coding_agents.md updated to reference the new
    layout.

Verification:
  - pytest tests/local_tests/ --collect-only: 62 tests collected, no
    import errors.
  - pytest tests/local_tests/wan22/ and stable_audio smoke test: pass.
  - lsp diagnostics on tests/local_tests/: 0 errors across 33 files.
