+++ tests/local_tests/kandinsky5/test_kandinsky5_lite_transformer_parity.py
