+++ tests/local_tests/ltx2/test_ltx2_gemma_encoder.py
