+++ tests/local_tests/ltx2/test_ltx2_pipeline_smoke.py
