+++ tests/local_tests/stable_audio/test_stable_audio_a2a_parity.py
@@ -15,7 +15,7 @@ import pytest
 import torch
 
 from fastvideo.utils import resolve_hf_token
-from tests.local_tests._stable_audio_helpers import (
+from ._stable_audio_helpers import (
     can_access_repo,
     setup_hf_env,
 )
