+++ tests/local_tests/stable_audio/test_stable_audio_pipeline_smoke.py
